代码改造完成，加入多字典查询方式
