+++ src/views/generator/generator.vue
@@ -1,6 +1,6 @@
 <template>
   <div>
-    <el-button type="primary" size="mini" @click="to">生成代码</el-button>
+    <el-button type="primary" size="mini" @click="toGen">生成代码</el-button>
     <el-dialog :visible.sync="dialog" title="代码生成配置" append-to-body width="800px">
       <el-table v-loading="loading" :data="data" size="small" style="width: 100%;">
         <el-table-column label="序号" width="80" align="center">
@@ -66,7 +66,7 @@ export default {
     }
   },
   methods: {
-    to() {
+    toGen() {
       this.dialog = true
       this.time = 130
       this.$nextTick(() => {
