--- src/views/monitor/log/module/header.vue
@@ -1,28 +1,0 @@
-<template>
-  <div class="head-container">
-    <el-input v-model="query.username" clearable placeholder="用户名" style="width: 200px;" class="filter-item" @keyup.enter.native="toQuery"/>
-    <el-button class="filter-item" size="mini" type="success" icon="el-icon-search" @click="toQuery">搜索</el-button>
-  </div>
-</template>
-
-<script>
-export default {
-  props: {
-    query: {
-      type: Object,
-      required: true
-    }
-  },
-  data() {
-    return {
-      downloadLoading: false
-    }
-  },
-  methods: {
-    toQuery() {
-      this.$parent.page = 0
-      this.$parent.init()
-    }
-  }
-}
-</script>
