+++ src/views/system/dept/index.vue
@@ -1,6 +1,34 @@
 <template>
   <div class="app-container">
-    <eHeader :query="query" :dicts="dicts"/>
+    <!--工具栏-->
+    <div class="head-container">
+      <!-- 搜索 -->
+      <el-input v-model="query.value" clearable placeholder="输入部门名称搜索" style="width: 200px;" class="filter-item" @keyup.enter.native="toQuery"/>
+      <el-select v-model="query.enabled" clearable placeholder="状态" class="filter-item" style="width: 90px" @change="toQuery">
+        <el-option v-for="item in enabledTypeOptions" :key="item.key" :label="item.display_name" :value="item.key"/>
+      </el-select>
+      <el-button class="filter-item" size="mini" type="success" icon="el-icon-search" @click="toQuery">搜索</el-button>
+      <!-- 新增 -->
+      <div v-permission="['ADMIN','DEPT_ALL','DEPT_CREATE']" style="display: inline-block;margin: 0px 2px;">
+        <el-button
+          class="filter-item"
+          size="mini"
+          type="primary"
+          icon="el-icon-plus"
+          @click="add">新增</el-button>
+      </div>
+      <div style="display: inline-block;">
+        <el-button
+          class="filter-item"
+          size="mini"
+          type="warning"
+          icon="el-icon-more"
+          @click="changeExpand">{{ $parent.expand ? '折叠' : '展开' }}</el-button>
+        <eForm ref="form" :is-add="true" :dicts="dicts"/>
+      </div>
+    </div>
+    <!--表单组件-->
+    <eForm ref="form" :is-add="isAdd" :dicts="dicts"/>
     <!--表格渲染-->
     <tree-table v-loading="loading" :expand-all="expand" :data="data" :columns="columns" size="small">
       <el-table-column label="状态" align="center">
@@ -17,7 +45,7 @@
       </el-table-column>
       <el-table-column v-if="checkPermission(['ADMIN','DEPT_ALL','DEPT_EDIT','DEPT_DELETE'])" label="操作" width="130px" align="center">
         <template slot-scope="scope">
-          <edit v-permission="['ADMIN','DEPT_ALL','DEPT_EDIT']" :dicts="dicts" :data="scope.row" :sup_this="sup_this"/>
+          <el-button v-permission="['ADMIN','DEPT_ALL','DEPT_EDIT']" size="mini" type="primary" icon="el-icon-edit" @click="edit(scope.row)"/>
           <el-popover
             v-permission="['ADMIN','DEPT_ALL','DEPT_DELETE']"
             :ref="scope.row.id"
@@ -43,10 +71,9 @@ import initData from '@/mixins/initData'
 import initDict from '@/mixins/initDict'
 import { del } from '@/api/dept'
 import { parseTime } from '@/utils/index'
-import eHeader from './module/header'
-import edit from './module/edit'
+import eForm from './form'
 export default {
-  components: { eHeader, edit, treeTable },
+  components: { eForm, treeTable },
   mixins: [initData, initDict],
   data() {
     return {
@@ -56,7 +83,11 @@ export default {
           value: 'name'
         }
       ],
-      delLoading: false, sup_this: this, expand: true
+      enabledTypeOptions: [
+        { key: 'true', display_name: '正常' },
+        { key: 'false', display_name: '禁用' }
+      ],
+      delLoading: false, expand: true
     }
   },
   created() {
@@ -96,6 +127,29 @@ export default {
         this.$refs[id].doClose()
         console.log(err.response.data.message)
       })
+    },
+    add() {
+      this.isAdd = true
+      const _this = this.$refs.form
+      _this.getDepts()
+      _this.dialog = true
+    },
+    changeExpand() {
+      this.expand = !this.expand
+      this.init()
+    },
+    edit(data) {
+      this.isAdd = false
+      const _this = this.$refs.form
+      _this.getDepts()
+      _this.form = {
+        id: data.id,
+        name: data.name,
+        pid: data.pid,
+        createTime: data.createTime,
+        enabled: data.enabled.toString()
+      }
+      _this.dialog = true
     }
   }
 }
