+++ src/views/system/dict/form.vue
@@ -1,14 +1,11 @@
 <template>
-  <el-dialog :visible.sync="dialog" :title="isAdd ? '新增权限' : '编辑权限'" append-to-body width="500px">
+  <el-dialog :append-to-body="true" :visible.sync="dialog" :title="isAdd ? '新增字典' : '编辑字典'" width="500px">
     <el-form ref="form" :model="form" :rules="rules" size="small" label-width="80px">
-      <el-form-item label="名称" prop="name">
-        <el-input v-model="form.name" style="width: 360px;"/>
+      <el-form-item label="字典名称" prop="name">
+        <el-input v-model="form.name" style="width: 370px;"/>
       </el-form-item>
-      <el-form-item label="别名" prop="alias">
-        <el-input v-model="form.alias" style="width: 360px;"/>
-      </el-form-item>
-      <el-form-item style="margin-bottom: 0px;" label="上级类目">
-        <treeselect v-model="form.pid" :options="permissions" style="width: 360px;" placeholder="选择上级类目" />
+      <el-form-item label="描述">
+        <el-input v-model="form.remark" style="width: 370px;"/>
       </el-form-item>
     </el-form>
     <div slot="footer" class="dialog-footer">
@@ -19,31 +16,25 @@
 </template>
 
 <script>
-import { add, edit, getPermissionTree } from '@/api/permission'
-import Treeselect from '@riophae/vue-treeselect'
-import '@riophae/vue-treeselect/dist/vue-treeselect.css'
+import { add, edit } from '@/api/dict'
 export default {
-  components: { Treeselect },
   props: {
     isAdd: {
       type: Boolean,
       required: true
-    },
-    sup_this: {
-      type: Object,
-      default: null
     }
   },
   data() {
     return {
-      loading: false, dialog: false, permissions: [],
-      form: { name: '', alias: '', pid: 0 },
+      loading: false, dialog: false,
+      form: {
+        id: '',
+        name: '',
+        remark: ''
+      },
       rules: {
         name: [
           { required: true, message: '请输入名称', trigger: 'blur' }
-        ],
-        alias: [
-          { required: true, message: '请输入别名', trigger: 'blur' }
         ]
       }
     }
@@ -59,8 +50,6 @@ export default {
           if (this.isAdd) {
             this.doAdd()
           } else this.doEdit()
-        } else {
-          return false
         }
       })
     },
@@ -73,7 +62,7 @@ export default {
           duration: 2500
         })
         this.loading = false
-        this.$parent.$parent.init()
+        this.$parent.init()
       }).catch(err => {
         this.loading = false
         console.log(err.response.data.message)
@@ -88,7 +77,7 @@ export default {
           duration: 2500
         })
         this.loading = false
-        this.sup_this.init()
+        this.$parent.init()
       }).catch(err => {
         this.loading = false
         console.log(err.response.data.message)
@@ -97,15 +86,11 @@ export default {
     resetForm() {
       this.dialog = false
       this.$refs['form'].resetFields()
-      this.form = { name: '', alias: '', pid: 0 }
-    },
-    getPermissions() {
-      getPermissionTree().then(res => {
-        this.permissions = []
-        const permission = { id: 0, label: '顶级类目', children: [] }
-        permission.children = res
-        this.permissions.push(permission)
-      })
+      this.form = {
+        id: '',
+        name: '',
+        remark: ''
+      }
     }
   }
 }
