--- src/views/system/dict/module/header.vue
@@ -1,46 +1,0 @@
-<template>
-  <div class="head-container">
-    <!-- 搜索 -->
-    <el-input v-model="query.value" clearable placeholder="输入搜索内容" style="width: 200px;" class="filter-item" @keyup.enter.native="toQuery"/>
-    <el-select v-model="query.type" clearable placeholder="类型" class="filter-item" style="width: 130px">
-      <el-option v-for="item in queryTypeOptions" :key="item.key" :label="item.display_name" :value="item.key"/>
-    </el-select>
-    <el-button class="filter-item" size="mini" type="success" icon="el-icon-search" @click="toQuery">搜索</el-button>
-    <div style="display: inline-block;margin: 0px 2px;">
-      <eForm ref="form" :is-add="true" :sup_this="sup_this"/>
-    </div>
-  </div>
-</template>
-
-<script>
-import checkPermission from '@/utils/permission' // 权限判断函数
-import eForm from './form'
-export default {
-  components: { eForm },
-  props: {
-    query: {
-      type: Object,
-      required: true
-    },
-    sup_this: {
-      type: Object,
-      required: true
-    }
-  },
-  data() {
-    return {
-      queryTypeOptions: [
-        { key: 'name', display_name: '字典名称' },
-        { key: 'remark', display_name: '描述' }
-      ]
-    }
-  },
-  methods: {
-    checkPermission,
-    toQuery() {
-      this.sup_this.page = 0
-      this.sup_this.init()
-    }
-  }
-}
-</script>
