+++ src/views/system/job/form.vue
@@ -33,10 +33,6 @@ export default {
       type: Boolean,
       required: true
     },
-    sup_this: {
-      type: Object,
-      default: null
-    },
     dicts: {
       type: Array,
       required: true
@@ -94,7 +90,7 @@ export default {
           duration: 2500
         })
         this.loading = false
-        this.$parent.$parent.init()
+        this.$parent.init()
       }).catch(err => {
         this.loading = false
         console.log(err.response.data.message)
@@ -109,7 +105,7 @@ export default {
           duration: 2500
         })
         this.loading = false
-        this.sup_this.init()
+        this.$parent.init()
       }).catch(err => {
         this.loading = false
         console.log(err.response.data.message)
