--- src/views/system/job/module/edit.vue
@@ -1,49 +1,0 @@
-<template>
-  <div>
-    <el-button size="mini" type="primary" icon="el-icon-edit" @click="to"/>
-    <eForm ref="form" :sup_this="sup_this" :is-add="false" :dicts="dicts"/>
-  </div>
-</template>
-<script>
-import eForm from './form'
-export default {
-  components: { eForm },
-  props: {
-    data: {
-      type: Object,
-      required: true
-    },
-    sup_this: {
-      type: Object,
-      required: true
-    },
-    dicts: {
-      type: Array,
-      required: true
-    }
-  },
-  methods: {
-    to() {
-      const _this = this.$refs.form
-      _this.getDepts()
-      _this.form = {
-        id: this.data.id,
-        name: this.data.name,
-        sort: this.data.sort,
-        enabled: this.data.enabled.toString(),
-        createTime: this.data.createTime,
-        dept: { id: this.data.dept.id }
-      }
-      _this.deptId = this.data.dept.id
-      _this.dialog = true
-    }
-  }
-}
-</script>
-
-<style scoped>
-  div{
-    display: inline-block;
-    margin-right: 3px;
-  }
-</style>
