+++ src/views/system/role/form.vue
@@ -41,10 +41,6 @@ export default {
     isAdd: {
       type: Boolean,
       required: true
-    },
-    sup_this: {
-      type: Object,
-      default: null
     }
   },
   data() {
@@ -99,7 +95,7 @@ export default {
           duration: 2500
         })
         this.loading = false
-        this.$parent.$parent.init()
+        this.$parent.init()
       }).catch(err => {
         this.loading = false
         console.log(err.response.data.message)
@@ -114,7 +110,7 @@ export default {
           duration: 2500
         })
         this.loading = false
-        this.sup_this.init()
+        this.$parent.init()
       }).catch(err => {
         this.loading = false
         console.log(err.response.data.message)
