--- src/views/system/user/center/log.vue
@@ -1,55 +1,0 @@
-<template>
-  <div>
-    <!--表格渲染-->
-    <el-table v-loading="loading" :data="data" size="small" style="width: 100%;">
-      <el-table-column prop="description" label="行为"/>
-      <el-table-column prop="requestIp" label="IP"/>
-      <el-table-column prop="time" label="请求耗时" align="center">
-        <template slot-scope="scope">
-          <el-tag v-if="scope.row.time <= 300">{{ scope.row.time }}ms</el-tag>
-          <el-tag v-else-if="scope.row.time <= 1000" type="warning">{{ scope.row.time }}ms</el-tag>
-          <el-tag v-else type="danger">{{ scope.row.time }}ms</el-tag>
-        </template>
-      </el-table-column>
-      <el-table-column prop="createTime" label="创建日期" width="180px">
-        <template slot-scope="scope">
-          <span>{{ parseTime(scope.row.createTime) }}</span>
-        </template>
-      </el-table-column>
-    </el-table>
-    <!--分页组件-->
-    <el-pagination
-      :total="total"
-      :current-page="page + 1"
-      style="margin-top: 8px;"
-      layout="total, prev, pager, next, sizes"
-      @size-change="sizeChange"
-      @current-change="pageChange"/>
-  </div>
-</template>
-
-<script>
-import initData from '@/mixins/initData'
-import { parseTime } from '@/utils/index'
-export default {
-  mixins: [initData],
-  created() {
-    this.$nextTick(() => {
-      this.init()
-    })
-  },
-  methods: {
-    parseTime,
-    beforeInit() {
-      this.url = 'api/logs/user'
-      const sort = 'id,desc'
-      this.params = { page: this.page, size: this.size, sort: sort }
-      return true
-    }
-  }
-}
-</script>
-
-<style scoped>
-
-</style>
